Feature/filters2

+++ project/app/fileSystemManager.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import re
 import sys
@@ -6,13 +7,14 @@ from app.models import *
 from scripts.esScript import es
 from app.utils import getRandomNumber, ImageFeature, processingLock
 
-
 class Node:
     def __init__(self, name: str, id: int, terminated=False):
         self.name = name  # folder name
         self.children = {}  # key: folder name, value: node
         self.parent = None
         self.id = id
+
+        self.deleting = False
 
         # if current node is the last one of an uri, it is terminated
         # ex: for uri C:\User\abcd\ef\, node 'ef' is terminated
@@ -49,7 +51,7 @@ class SimpleFileSystemManager:
         else:
             return False
 
-        return node.terminated
+        return node.terminated ^ node.deleting
 
     def add_full_path_uri(self, uri: str, ids: list):
         folders = re.split("[\\\/]+", uri)
@@ -156,75 +158,86 @@ class SimpleFileSystemManager:
             path = os.path.join(path, p)
         return path
 
-    def delete_folder_from_fs(self, uri):
-        if self.exist(uri):
-            node = self.get_last_node(uri)
+    def delete_folder_from_fs(self, uri, frr):
+        try:
+            if self.exist(uri):
+                node = self.get_last_node(uri)
 
-            if node and node.parent:
-                node.parent.delete_node(node)
-
-            foldersto_be_deleted = [Folder.nodes.get_or_none(id_=node.id)]
-
-            deletedImages = []
-
-            while foldersto_be_deleted != []:
-                f = foldersto_be_deleted.pop()
-                if not f:
-                    continue
-
-                images = f.getImages()
-
-                if images is not None:
-                    for image in images:
-                        if len(image.folder) > 1: # if image is in different folders
-                            currentImageUri, root = self.__splitUriAndGetRoot__(image.folder_uri)
-                            currentImageUri = self.__builFullPath__(currentImageUri)
-                            if currentImageUri == self.__fullPathForFolderNode__(f):
-                                for folder in image.folder:
-                                    if folder.id != f.id and not folder.isChildOf(f.id_):
-                                        image.folder_uri = self.__fullPathForFolderNode__(folder)
-                                        image.save()
-                                        es_image = ImageES.get(using=es, index='image', id=image.hash)
-                                        es_image.uri = image.folder_uri
-                                        es_image.save(using=es)
-                                        break
-                        else:
-                            self.delete_connected_tags_and_persons(image)
-                            self.delete_locations(image)
-                            image.delete()
-
-                            thumbnail = os.path.join("app", "static", "thumbnails", str(image.hash)) + ".jpg"
-                            os.remove(thumbnail)
-
-                            es_image = ImageES.get(using=es, index='image', id=image.hash)
-                            es_image.delete(using=es)
-                            deletedImages.append(ImageFeature(hash=image.hash))
-
-                childrenFolders = f.getChildren()
-                if childrenFolders:
-                    foldersto_be_deleted.extend(list(childrenFolders))
-                f.delete()
-
-            # if parent folders have no children folders and no images, delete it
-            while node.parent:
-                parent_folder = Folder.nodes.get_or_none(id_=node.parent.id)
-                if not parent_folder:
+                if node and node.parent:
                     node.parent.delete_node(node)
+
+                node.deleting = True
+
+                foldersto_be_deleted = [Folder.nodes.get_or_none(id_=node.id)]
+
+                deletedImages = []
+
+                while foldersto_be_deleted != []:
+                    f = foldersto_be_deleted.pop()
+                    if not f:
+                        continue
+
+                    images = f.getImages()
+
+                    if images is not None:
+                        for image in images:
+                            frr.remove_image(image.hash)
+                            if len(image.folder) > 1: # if image is in different folders
+                                currentImageUri, root = self.__splitUriAndGetRoot__(image.folder_uri)
+                                currentImageUri = self.__builFullPath__(currentImageUri)
+                                if currentImageUri == self.__fullPathForFolderNode__(f):
+                                    for folder in image.folder:
+                                        if folder.id != f.id and not folder.isChildOf(f.id_):
+                                            image.folder_uri = self.__fullPathForFolderNode__(folder)
+                                            image.save()
+                                            es_image = ImageES.get(using=es, index='image', id=image.hash)
+                                            es_image.uri = image.folder_uri
+                                            es_image.save(using=es)
+                                            break
+                            else:
+                                self.delete_connected_tags_and_persons(image)
+                                self.delete_locations(image)
+                                image.delete()
+
+                                try:
+                                    thumbnail = os.path.join("app", "static", "thumbnails", str(image.hash)) + ".jpg"
+                                    os.remove(thumbnail)
+
+                                    es_image = ImageES.get(using=es, index='image', id=image.hash)
+                                    es_image.delete(using=es)
+                                    deletedImages.append(ImageFeature(hash=image.hash))
+                                except Exception as e:
+                                    logging.info("[Deleting]: [ERROR] Image missing: " + e)
+
+
+                    childrenFolders = f.getChildren()
+                    if childrenFolders:
+                        foldersto_be_deleted.extend(list(childrenFolders))
+                    f.delete()
+
+                # if parent folders have no children folders and no images, delete it
+                while node.parent:
+                    parent_folder = Folder.nodes.get_or_none(id_=node.parent.id)
+                    if not parent_folder:
+                        node.parent.delete_node(node)
+                        node = node.parent
+                        continue
+                    curr_node = node
                     node = node.parent
-                    continue
-                curr_node = node
-                node = node.parent
-                children = parent_folder.getChildren()
-                if len(children) == 0 and not parent_folder.terminated:
-                    parent_folder.delete()
-                    if curr_node.name in node.children:
-                        node.children.pop(curr_node.name)
-                else: break
+                    children = parent_folder.getChildren()
+                    if len(children) == 0 and not parent_folder.terminated:
+                        parent_folder.delete()
+                        if curr_node.name in node.children:
+                            node.children.pop(curr_node.name)
+                    else: break
 
-            if node.name in self.trees and len(node.children) == 0:
-                self.trees.pop(node.name)
+                if node.name in self.trees and len(node.children) == 0:
+                    self.trees.pop(node.name)
 
-            return deletedImages
+                return deletedImages
+        except Exception as e:
+            logging.info("[Deleting]: [ERROR] Delete error " + e)
+
 
 
     def __fullPathForFolderNode__(self, f):
@@ -234,15 +247,17 @@ class SimpleFileSystemManager:
         return self.__builFullPath__(paths)
 
     def delete_connected_tags_and_persons(self, image):
-        for t in image.tag:
-            t.image.disconnect(image)
-            if len(t.image) == 0:
-                t.delete()
+        if len(image.tag) != 0:
+            for t in image.tag:
+                t.image.disconnect(image)
+                if len(t.image) == 0:
+                    t.delete()
 
-        for p in image.person:
-            p.image.disconnect(image)
-            if len(p.image) == 0:
-                p.delete()
+        if len(image.person) != 0:
+            for p in image.person:
+                p.image.disconnect(image)
+                if len(p.image) == 0:
+                    p.delete()
 
     def delete_locations(self, image):
         if len(image.location) != 0:
